use loadEsm
